When Use_CGAL is ON: create an OpenMEEGCgal library, so that meshing tools link to that lib.
