+++ OpenMEEG/tools/CMakeLists.txt
@@ -38,34 +38,40 @@ add_executable(om_mesh_to_dip mesh_to_dip.cpp)
 target_link_libraries(om_mesh_to_dip OpenMEEG ${VTK_LIBRARIES})
 
 if (USE_CGAL)
-    set(OM_CGAL_TARGETS cgal_mesh_function cgal_remesh)
+    # compile the OpenMEEGCgal library
+    add_library(OpenMEEGCgal cgal_mesh.cpp)
+    target_link_libraries(OpenMEEGCgal OpenMEEG ${CGAL_LIBRARIES} ${VTK_LIBRARIES})
+    target_include_directories(OpenMEEGCgal PUBLIC ${CGAL_INCLUDE_DIRS})
+    target_compile_options(OpenMEEGCgal PUBLIC ${CGAL_CXX_FLAGS})
+
+    set(OM_CGAL_TARGETS cgal_mesh_function)
     set(OM_CGAL_EXEC)
 
     if (CGAL_VERSION VERSION_GREATER 4.6)
         # edge collapse feature is stabilized in version ? not 4.2
-        list(APPEND OM_CGAL_TARGETS "cgal_decimate")
+        list(APPEND OM_CGAL_TARGETS "cgal_decimate" "cgal_remesh")
     endif()
 
     foreach(target ${OM_CGAL_TARGETS})
         add_executable(om_${target} ${target}.cpp)
         list(APPEND OM_CGAL_EXEC om_${target})
-        target_link_libraries(om_${target} OpenMEEG ${CGAL_LIBRARIES} ${VTK_LIBRARIES})
+        target_link_libraries(om_${target} OpenMEEG OpenMEEGCgal ${CGAL_LIBRARIES} ${VTK_LIBRARIES})
         target_include_directories(om_${target} PUBLIC ${CGAL_INCLUDE_DIRS})
         target_compile_options(om_${target} PUBLIC ${CGAL_CXX_FLAGS})
     endforeach()
-    if (CGAL_ImageIO_FOUND) # these are tools hanling images (.inr format only for the moment!)
+    if (GAL_ImageIO_FOUND) # these are tools hanling images (.inr format only for the moment!)
         foreach(target cgal_mesh_3Dlevelset_image cgal_mesh_3Dlabeled_image)
             add_executable(om_${target} ${target}.cpp)
             list(APPEND OM_CGAL_EXEC om_${target})
-            target_link_libraries(om_${target} OpenMEEG ${CGAL_LIBRARIES} ${VTK_LIBRARIES})
+            target_link_libraries(om_${target} OpenMEEG OpenMEEGCgal ${CGAL_LIBRARIES} ${VTK_LIBRARIES})
             target_include_directories(om_${target} PUBLIC ${CGAL_INCLUDE_DIRS})
             target_compile_options(om_${target} PUBLIC ${CGAL_CXX_FLAGS})
         endforeach()
     endif()
     install(TARGETS ${OM_CGAL_EXEC} DESTINATION bin)
     if (BUILD_TESTING)
-        OPENMEEG_TEST(cgal_remesh om_cgal_remesh -i ${CMAKE_SOURCE_DIR}/data/Head1/cortex.1.tri -o ${OpenMEEG_BINARY_DIR}/tests/remeshed.tri)
         if (CGAL_VERSION VERSION_GREATER 4.6)
+            OPENMEEG_TEST(cgal_remesh om_cgal_remesh -i ${CMAKE_SOURCE_DIR}/data/Head1/cortex.1.tri -o ${OpenMEEG_BINARY_DIR}/tests/remeshed.tri)
             OPENMEEG_TEST(cgal_decimate om_cgal_decimate -i ${CMAKE_SOURCE_DIR}/data/Head3/cortex.3.tri -n 100 -o ${OpenMEEG_BINARY_DIR}/tests/decimated.tri)
         endif()
     endif()
